:zap: Created HOC :)
This allowed me to drastically reduce duplication. This is a very very good pattern.
