+++ src/components/PianoRoll.vue
@@ -52,91 +52,7 @@ export default class PianoRoll extends Vue {
   @Inject() public pxPerBeat!: number;
   @Prop({ type: Object, required: true }) public instrument!: Instrument;
   @Prop({ type: Array, required: true }) public value!: Note[];
-  @Prop({ type: Boolean, required: true }) public play!: boolean;
   @Prop({ type: Object, required: true }) public part!: Part<Note>;
-
-  public loopStart = 0;
-  public loopEnd = 0;
-  public scrollLeft = 0;
-  public progress = 0;
-  public sequencerLoopEnd = 0;
-  public setLoopStart: null | number = null;
-  public setLoopEnd: null | number = null;
-
-  public update() {
-    if (this.part.state === 'started') { requestAnimationFrame(this.update); }
-    this.progress = this.part.progress;
-  }
-
-  public added(note: Note) {
-    // TODO There is duplication here with project.ts
-    // Eventually, we will need a solution.
-    const time = toTickTime(note.time);
-    this.$log.debug(`Adding note at ${note.time} -> ${time}`);
-    const callback = this.instrument.callback.bind(this.instrument);
-    this.part.add(callback, time, note);
-  }
-
-  public removed(note: Note, i: number) {
-    this.part.remove(note);
-  }
-
-  public scrollHorizontal(scrollLeft: number) {
-    this.scrollLeft = scrollLeft;
-  }
-
-  // Horizontal offset in beats.
-  public get offset() {
-    return this.scrollLeft / this.pxPerBeat;
-  }
-
-  @Watch<PianoRoll>('part')
-  public resetLoop() {
-    this.setLoopStart = null;
-    this.setLoopEnd = null;
-  }
-
-  @Watch<PianoRoll>('setLoopStart')
-  public changeLoopStart() {
-    this.loopStart = this.setLoopStart || 0;
-  }
-
-  @Watch<PianoRoll>('setLoopEnd')
-  public changeLoopEnd() {
-    if (this.setLoopEnd) {
-      this.loopEnd = this.setLoopEnd;
-    } else {
-      this.loopEnd = this.sequencerLoopEnd;
-    }
-  }
-
-  @Watch<PianoRoll>('sequencerLoopEnd', { immediate: true })
-  public changeSeqLoopEnd() {
-    if (!this.setLoopEnd) {
-      this.loopEnd = this.sequencerLoopEnd;
-    }
-  }
-
-  @Watch<PianoRoll>('play', { immediate: true })
-  public onPlay() {
-    if (this.play) {
-      this.update();
-    }
-  }
-
-  @Watch<PianoRoll>('loopEnd', { immediate: true })
-  public onLoopEndChange() {
-    this.$log.debug(`loodEnd being set to ${this.loopEnd}`);
-    this.part.loopEnd = toTickTime(this.loopEnd);
-  }
-
-  @Watch<PianoRoll>('loopStart', { immediate: true })
-  public onLoopStartChange() {
-    this.$log.debug(`loopStart being set to ${this.loopStart}`);
-    const time = toTickTime(this.loopStart);
-    this.part.seconds = new Tone.Time(time).toSeconds();
-    this.part.loopStart = time;
-  }
 }
 </script>
